requirements.txt
python-telegram-bot>=21.0
aiofiles>=23.0
python-dotenv>=1.0
